chore: change when to request notification permission
